vehicle/viss/0200-subscribe-unique-id-success.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>
[0200: unique subscriptionId]
</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">
<link rel="stylesheet" type="text/css" href="./style-general.css">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-config.js"></script>
<script src="./vehicle-test-helper.js"></script>

<div id="test-title">
[0200: unique subscriptionId]
</div>
<br>
<div id="test-assertion">
<b>Test assertion:</b><br>
The server ensures that a new unique subscription id value is returned<br>
 for each successful subscription request for a particular WebSocket connection.<br>
</div>
<div id="scenario">
<br>
<b>Test Scenario:</b><br>
1. Execute 'subscribe' method X times.<br>
   (e.g. path = 'Signal.Drivetrain.Transmission.Speed')<br>
2. Confirm 'SubscribeSuccessResponse' is returned for X times and<br>
   X number of subscriptionId is obtained.<br>
3. Confirm that all the obtained subscriptionId are unique and
   there is no set of subscriptionIds those have same value.
4. Execute 'unsubscribeAll' to unsubscribe all the subscriptions<br>
<br>
<b>Expected Result:</b><br>
- There is no duplicated subscriptionId<br>
</div>

<div id="result"></div>
<br>
<div id="log"></div>

<script>
var idnum_base = 200;
var subscribe_num = 10;

var reqId = [];
var subId = [];

// init arrays
for (var i=0; i<subscribe_num; i++) {
  reqId[i] = "reqid-0" + idnum_base+i;
  subId[i] = null;
}

var path = SUBSCRIBE_STANDARD_PATH;

// phase
var PH_START            = 0;
var PH_SENT_SUBSCRIBE   = 1;
var PH_SUBSCRIBED       = 2;
var PH_SENT_UNSUBSCRIBE = 3;
var PH_UNSUBSCRIBED     = 4;
var PH_END              = 5;

var phase = PH_START;

var t = async_test("Subscribe test: Unique subscriptionID");
var vehicle  = new WebSocket(VISS_URL, VISS_SUBPROTO);

vehicle.onopen = function() {

  // process messages from the server
  vehicle.onmessage = function(event) {
    addLogMessage("response: json=" + event.data);
    var msg = JSON.parse(event.data);

    if (phase == PH_START) {
      // nothing to do
    } else if (phase == PH_SENT_SUBSCRIBE) {
      t.step_func(function() {
        // wait for subscribe response
        if (msg.action !== "subscribe") {
          addLogMessage("PH_SENT_SUBSCRIBE: Ignore messages except for Subscribe response.");
          return;
        }
        var res = isSubscribeSuccessResponse("", msg);
        assert_true(res, "True if subscribeSuccessResponse received.");
        if (res) {
          for (var i=0; i<subscribe_num; i++) {
            if (msg.requestId === reqId[i]) {
              subId[i] = msg.subscriptionId;
              addLogMessage("subscribeSuccessResponse for reqId["+i+"] received. subscribeId = "
                            + subId[i]);
            }
          }
        } else {
          helper_terminate_failure("subscribeErrorResponse received. test failed.");
          unsubscribeAllRequest(subId);
          phase = PH_END;
        }

        // check if all the subscribe() response received
        var passed = subId.every(function(element, index, array) { return (element != null); });

        if (passed) {
          //all subscribe() received subscribeSuccessResponse
          phase = PH_SUBSCRIBED;
          unsubscribeAllRequest(subId);

          var subId0 = subId[0];
          var subIdRest = subId.slice(1);

          addLogMessage("<br>Confirm there is no duplicated subscriptionId.");
          // check if no subId is duplicate of subId0
          var success = subIdRest.every(function(element, index, array) {
            var idx = index + 1;
            if (element === subId0) {
              addLogMessage("subId["+ idx +"] == subId[0]");
              return false;
            } else {
              addLogMessage("subId["+ idx +"] != subId[0]");
              return true;
            }
          });
          if (success) {
            addLogMessage("Test Success: There is no duplicated subId of subId[0] = " + subId[0] + "<br>");
            helper_terminate_success("There is no duplicated of subId[0] = " + subId[0]);
          } else {
            addLogMessage("Test Failure: There is duplicated subId. Test Failed.<br>");
            helper_terminate_failure("There is duplicated subId = " + subId[0] + ". Test Failed.");
          }
          phase = PH_END;
        }

      })();
    } else if (phase == PH_SUBSCRIBED) {
    } else if (phase == PH_SENT_UNSUBSCRIBE) {
    } else if (phase == PH_UNSUBSCRIBED) {
    }
  }

  // in case server doesn't return response
  t.step_timeout(function() {
    if (phase < PH_SUBSCRIBED) {
      helper_terminate_failure("Finish by timeout");
    }
  }, TIME_OUT_TIME);

  addLogMessage("Send subscribe request: path="+path);
  phase = PH_SENT_SUBSCRIBE;
  var sub_request;
  for (var i=0; i< subscribe_num; i++) {
    sub_request = '{"action":"subscribe","path":"'+ path +'" ,"requestId":"'+ reqId[i] +'"}';
    addLogMessage("Sub request: json=" + sub_request);
    vehicle.send(sub_request);
  }
}

function unsubscribeAllRequest(_subId) {
  var reqId_base = 1000;
  var unsuball_request = '{"action":"unsubscribeAll","requestId":"'+reqId_base+'"}';
  addLogMessage("unsubAll request: json="+unsuball_request );
  vehicle.send( unsuball_request );
}

</script>
</html>
